+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project (not yet) adheres to [Semantic Versioning](https://semver.org/spec/
 
 ### Changed
 
+- Make tile_image field optional in DiPlan API - Bplans sent without an image automatically receive the placeholder image
 - Password visibility toggle added to login and signup forms
 - External projects/plans show icon to indicate external link
 - Offline Event is now A Module